+++ node/CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.  The format
 * Individual weights for traffic throttling can now be set through the configuration value
   `network.estimator_weights`.
 * Added `consensus.highway.max_request_batch_size` configuration parameter. Defaults to 20.
+* New histogram metrics `deploy_acceptor_accepted_deploy` and `deploy_acceptor_rejected_deploy` that track how long the initial verification took.
 
 ### Changed
 * The following Highway timers are now separate, configurable, and optional (if the entry is not in the config, the timer is never called):